+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 # v2.0.0
+**API changes**
+* generate.zoom_video_frames and generate.video_frames have been combined to a single function: generate.video_frames. If you do not specify zoom_scale, shift_x, or shift_y, these values default to 0, and non-zooming images are generated.
+* generate.video_frames arguments changed. iterations and save_every are removed. New arguments are provided to make it easier to calculate video durations.
+  * num_video_frames : Set the number of video frames (images) to be generated.
+  * iterations_per_frame : Set the number of vqgan training iterations to perform for each frame of video. Higher numbers are more stylized.
+* 
+
+**New Features**
+* generate.zoom_video lets you specify specific video frames where prompts should be changed using the argument change_prompts_on_frame. E.g. to change prompts on frames 150 and 200, use change_prompts_on_frame = [150,200]. Examples are updated with this argument.
+* video_tools now sets ffmpeg to output on error only
 
 **Bug Fixes**
 * upscaling video example file had a bug in the ffmpeg command. Fixed.